requirements.txt
asgiref==3.3.1
attrs==20.3.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-import-export==2.5.0
djangorestframework==3.12.2
django-rest-registration
et-xmlfile==1.1.0
gunicorn==20.1.0
importlib-metadata==3.10.0
jsonschema==3.2.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pyrsistent==0.17.3
python-dotenv==1.0.0
pytz==2021.1
PyYAML==6.0.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.7.4.3
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
